requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
brotli==1.0.9
    # via whitenoise
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
django==4.1.3
    # via
    #   -r requirements.in
    #   django-import-export
    #   django-sesame
django-environ==0.9.0
    # via -r requirements.in
django-import-export==3.0.1
    # via -r requirements.in
django-sesame[ua]==3.1
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements.in
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.10
    # via tablib
psycopg2-binary==2.9.5
    # via -r requirements.in
pymemcache==4.0.0
    # via -r requirements.in
pyyaml==6.0
    # via tablib
sqlparse==0.4.3
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.1
    # via django-import-export
ua-parser==0.16.1
    # via django-sesame
whitenoise[brotli]==6.2.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
